#4 New "edit profile" page, user slugs + endpoint
