+++ src/components/Header/Header.js
@@ -1,9 +1,16 @@
+/* eslint-disable max-len */
 import { useSession } from 'next-auth/client';
 import { useRouter } from 'next/router';
 import Container from 'react-bootstrap/Container';
 import { Nav, Navbar, NavDropdown } from 'react-bootstrap';
 import { BoxArrowUpRight, BoxArrowInRight, BoxArrowRight } from 'react-bootstrap-icons';
 import SecondNavbar from '../SecondNavbar';
+
+function getEditProfileUrl(email) {
+  const slug = email.replace(/[*+~.()'"!:@]/g, '-');
+  const editUrl = `/user/${slug}/editprofile`;
+  return editUrl;
+}
 
 function Header() {
   const [session] = useSession();
@@ -33,15 +40,15 @@ function Header() {
                 <>
                   {session.user.name && (
                     <NavDropdown title={session.user.name} id="basic-nav-dropdown" data-testid="nav-profile-dropdown">
-                      <NavDropdown.Item href="#editprofile">My Profile</NavDropdown.Item>
+                      <NavDropdown.Item href={getEditProfileUrl(session.user.email)}>My Profile</NavDropdown.Item>
                       <NavDropdown.Item href="/api/auth/signout" data-testid="nav-login-link">
                         Log Out
                         <BoxArrowRight className="align-text-bottom ml-1" />
                       </NavDropdown.Item>
                     </NavDropdown>
                   )}
-                  {!session.user.name && router.pathname !== '/auth/newuser' && (
-                    <Nav.Link href="/auth/newuser" className="text-danger">
+                  {!session.user.name && router.pathname !== '/user/newuser' && (
+                    <Nav.Link href="/user/newuser" className="text-danger">
                       Complete Registration
                     </Nav.Link>
                   )}
