+++ src/pages/api/user/slug/[slug].js
@@ -1,5 +1,5 @@
 import nc from 'next-connect';
-import middleware from '../../../middlewares/middleware';
+import middleware from '../../../../middlewares/middleware';
 
 
 const handler = nc()
@@ -9,15 +9,15 @@ const handler = nc()
       await req.db
         .collection('users')
         .findOne(
-          { email: req.query.email },
+          { slug: req.query.slug },
           (err, doc) => {
             if (doc) {
               const {
-                name, firstName, lastName, affiliation,
+                name, firstName, lastName, affiliation, email,
               } = doc;
               if (err) throw err;
               res.status(200).json({
-                name, firstName, lastName, affiliation,
+                name, firstName, lastName, affiliation, email,
               });
             } else {
               res.status(404).json({});
