+++ src/pages/user/newuser.jsx
@@ -6,34 +6,16 @@ import * as yup from 'yup';
 import {
   Button, Card, Col, Form, Row, Spinner,
 } from 'react-bootstrap';
+import Feedback from 'react-bootstrap/Feedback';
+import Link from 'next/link';
 import Router from 'next/router';
 import fullName from '../../utils/nameUtil';
 import Layout from '../../components/Layout';
 
-const EditProfile = () => {
+const NewUser = () => {
   const [session] = useSession();
 
   const [errorMsg, setErrorMsg] = useState('');
-
-  const getInitialValues = async () => {
-    const { email } = session.user;
-    const url = `/api/user/${email}`;
-    const res = await fetch(url, {
-      method: 'GET',
-      headers: { 'Content-Type': 'application/json' },
-    });
-    if (res.status === 200) {
-      const user = await res.json();
-      setErrorMsg(JSON.stringify(user));
-      const { firstName, lastName, affiliation } = user;
-      return Promise.resolve({
-        email, firstName, lastName, affiliation,
-      });
-    }
-    return Promise.resolve({
-      error: 'User not found',
-    });
-  };
 
   const submitHandler = async (values) => {
     const body = {
@@ -42,6 +24,7 @@ const EditProfile = () => {
       lastName: values.lastName,
       name: fullName(values.firstName, values.lastName),
       affiliation: values.affiliation,
+      slug: session.user.email.replace(/[*+~.()'"!:@]/g, '-'),
     };
     const res = await fetch('/api/users', {
       method: 'PATCH',
@@ -53,9 +36,7 @@ const EditProfile = () => {
       getSession();
       Router.push({
         pathname: '/',
-        query: {
-          alert: 'profileEdited',
-        },
+        query: { alert: 'completeRegistration' },
       });
     } else {
       setErrorMsg(await res.text());
@@ -66,6 +47,9 @@ const EditProfile = () => {
     firstName: yup.string().required('Required'),
     lastName: yup.string().required('Required'),
     affiliation: yup.string().required('Required'),
+    tosCheck: yup.boolean()
+      .required('You must agree to the Terms and Conditions before registering.')
+      .oneOf([true], 'You must agree to the Terms and Conditions before registering.'),
   });
 
   return (
@@ -81,7 +65,11 @@ const EditProfile = () => {
           )}
           {session && (
             <Card.Body>
-              <Card.Title>Edit Profile</Card.Title>
+              <Card.Title>Welcome to Annotation Studio</Card.Title>
+              <Card.Text>
+                Please fill out the following form to complete your registration.
+              </Card.Text>
+
               <Formik
                 onSubmit={(values, actions) => {
                   setTimeout(() => {
@@ -90,8 +78,12 @@ const EditProfile = () => {
                   }, 1000);
                 }}
                 validationSchema={schema}
-                initialValues={getInitialValues}
-                enableReinitialize={true}
+                initialValues={{
+                  firstName: '',
+                  lastName: '',
+                  affiliation: '',
+                  tosCheck: false,
+                }}
               >
                 {(props) => (
                   <Form onSubmit={props.handleSubmit} noValidate>
@@ -117,7 +109,6 @@ const EditProfile = () => {
                           onChange={props.handleChange}
                           onBlur={props.handleBlur}
                           value={props.values.firstName}
-                          defaultValue={props.values.firstName}
                           isValid={props.touched.firstName && !props.errors.firstName}
                           isInvalid={!!props.errors.firstName}
                         />
@@ -165,6 +156,29 @@ const EditProfile = () => {
                         />
                       </Col>
                     </Form.Group>
+                    <Form.Group>
+                      <Form.Check type="checkbox">
+                        <Form.Check.Input
+                          required
+                          type="checkbox"
+                          name="tosCheck"
+                          id="tosCheck"
+                          value={props.values.tosCheck}
+                          onChange={props.handleChange}
+                          onBlur={props.handleBlur}
+                          isValid={props.touched.tosCheck && !props.errors.tosCheck}
+                          isInvalid={!!props.errors.tosCheck}
+                        />
+                        <Form.Check.Label htmlFor="tosCheck">
+                          I agree to the Annotation Studio
+                          {' '}
+                          <Link href="#tos"><a href="#tos" title="tos">Terms and Conditions</a></Link>
+                        </Form.Check.Label>
+                        <Feedback type="invalid">
+                          {props.errors.tosCheck}
+                        </Feedback>
+                      </Form.Check>
+                    </Form.Group>
                     <Row>
                       <Col className="text-right">
                         <Button
@@ -187,4 +201,4 @@ const EditProfile = () => {
   );
 };
 
-export default EditProfile;
+export default NewUser;
